refactor(serve): split chat_completions endpoint into a package
Aligns with the responses/ package layout. Splits the 633-line
chat_completions.py into protocol/validation/logprobs/logits_processors/
serving modules. Chat-specific models move to
endpoints/chat_completions/protocol.py; shared models stay in the
top-level protocol.py with backward-compat re-exports. No behavior change.
