+++ lmdeploy/serve/openai/endpoints/__init__.py
@@ -1,5 +1,6 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 """OpenAI and LMDeploy endpoint router assembly."""
+from __future__ import annotations
 
 from .router import create_openai_router
